Remove outdated Camel devfiles
* Camel K devfiles is no more working
* Camel on Springboot still relies on a very old Camel version which is
no more community supported.

There are nobody with time allocated to maintain these examples. They
would benefit from drastic changes given new orientations on Camel. So
removing for now and when we find people with time to work on it we can
re-add some later.
There is also mention of Fuse which is the product and that will be EOL
soon.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>